+++ pages/projects/virtual-assistant/index.js
@@ -3,7 +3,10 @@ import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { useTranslation } from "next-i18next";
 import { Layout } from "../../../components/organisms/Layout";
 import { ActionButton } from "../../../components//atoms/ActionButton";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
+import queryGraphQL from "../../../graphql/client";
+import getVirtualAssistantPage from "../../../graphql/queries/virtualAssistantQuery.graphql";
+import Image from "next/image";
 
 //  On hold for now
 //  import { VirtualConcierge } from "../../../components/organisms/VirtualConcierge";
@@ -11,7 +14,7 @@ import { useEffect } from "react";
 
 export default function Home(props) {
   const { t } = useTranslation(["common", "vc"]);
-  const language = props.locale === "en" ? "fr" : "en";
+  const [pageData] = useState(props.pageData.item);
 
   useEffect(() => {
     if (process.env.NEXT_PUBLIC_ADOBE_ANALYTICS_URL) {
@@ -24,7 +27,9 @@ export default function Home(props) {
     <>
       <Layout
         locale={props.locale}
-        langUrl={t("virtualAssistantPath")}
+        langUrl={
+          props.locale === "en" ? pageData.scPageNameFr : pageData.scPageNameEn
+        }
         breadcrumbItems={[
           { text: t("siteTitle"), link: t("breadCrumbsHref1") },
           { text: t("menuLink1"), link: t("breadCrumbsHref2") },
@@ -38,9 +43,9 @@ export default function Home(props) {
           )}
 
           {/* Primary HTML Meta Tags */}
-          <title>{`${t("vc:virtualAssistantTitle")} — ${t(
-            "siteTitle"
-          )}`}</title>
+          <title>{`${
+            props.locale === "en" ? pageData.scTitleEn : pageData.scTitleFr
+          } — ${t("siteTitle")}`}</title>
           <meta name="description" content={`${t("vc:metaDescription")}`} />
           <meta name="author" content="Service Canada" />
           <link rel="icon" href="/favicon.ico" />
@@ -109,31 +114,67 @@ export default function Home(props) {
         {/* Virtual Assitant Demo section start -  with link to working prototype */}
         <section className="layout-container mb-10">
           <h1 className="mb-8 text-h1l" tabIndex="-1" id="pageMainTitle">
-            {t("vc:virtualAssistantTitle")}
+            {props.locale === "en" ? pageData.scTitleEn : pageData.scTitleFr}
           </h1>
           {/* the scenario section */}
-          <div className="whitespace-pre-line ">
+          <div>
             <h2 className="mb-6 mt-8 text-h1" id="virtualAssistantScenario">
-              {t("vc:scenarioTitle")}
+              {props.locale === "en"
+                ? pageData.scFragments[0].scContentEn.json[0].content[0].value
+                : pageData.scFragments[0].scContentFr.json[0].content[0].value}
             </h2>
-            <p
-              className="mb-6 lg:col-span-2 px-1 lg:px-0 xl:w-3/4"
-              dangerouslySetInnerHTML={{ __html: t("vc:scenarioBody") }}
-            >
-              {/* {t("vc:scenarioDemoBody")} */}
+            <p className="mb-6 lg:col-span-2 px-1 lg:px-0 xl:w-3/4">
+              {props.locale === "en"
+                ? pageData.scFragments[0].scContentEn.json[1].content[0].value
+                : pageData.scFragments[0].scContentFr.json[1].content[0].value}
+              <strong>
+                {props.locale === "en"
+                  ? pageData.scFragments[0].scContentEn.json[1].content[1].value
+                  : pageData.scFragments[0].scContentFr.json[1].content[1]
+                      .value}
+              </strong>
+              {props.locale === "en"
+                ? pageData.scFragments[0].scContentEn.json[1].content[2].value
+                : pageData.scFragments[0].scContentFr.json[1].content[2].value}
+              <strong>
+                {props.locale === "en"
+                  ? pageData.scFragments[0].scContentEn.json[1].content[3].value
+                  : pageData.scFragments[0].scContentFr.json[1].content[3]
+                      .value}
+              </strong>
+              {props.locale === "en"
+                ? pageData.scFragments[0].scContentEn.json[1].content[4].value
+                : pageData.scFragments[0].scContentFr.json[1].content[4].value}
             </p>
+            <p className="mb-6 lg:col-span-2 px-1 lg:px-0 xl:w-3/4">
+              {props.locale === "en"
+                ? pageData.scFragments[0].scContentEn.json[2].content[0].value
+                : pageData.scFragments[0].scContentFr.json[2].content[0].value}
+              <strong>
+                {props.locale === "en"
+                  ? pageData.scFragments[0].scContentEn.json[2].content[1].value
+                  : pageData.scFragments[0].scContentFr.json[2].content[1]
+                      .value}
+              </strong>
+              {props.locale === "en"
+                ? pageData.scFragments[0].scContentEn.json[2].content[2].value
+                : pageData.scFragments[0].scContentFr.json[2].content[2].value}
+            </p>
+
             <p className="flex mb-16 text-center">
               <ActionButton
                 href={
-                  language === "en"
-                    ? "https://av-va.alpha.service.canada.ca/fr"
-                    : "https://av-va.alpha.service.canada.ca/en"
+                  props.locale === "en"
+                    ? pageData.scFragments[3].scDestinationURLEn
+                    : pageData.scFragments[3].scDestinationURLFr
                 }
                 id="meet-va-link"
                 dataCy="meet-va-link"
                 className="rounded px-6 py-4 font-bold text-center inline-block"
               >
-                {t("vc:meetAssistant")}
+                {props.locale === "en"
+                  ? pageData.scFragments[3].scTitleEn
+                  : pageData.scFragments[3].scTitleFr}
               </ActionButton>
               {/* href="https://av-va.alpha.service.canada.ca" */}
             </p>
@@ -142,18 +183,33 @@ export default function Home(props) {
           <div className="w-auto mb-6 ">
             <div className="flex flex-col break-words lg:grid lg:grid-cols-2 gap-4 lg:gap-6 ">
               <h2 className="mb-0 text-h1" id="virtualAssistantTitle">
-                {t("vc:virtualAssistantBioTitle")}
+                {props.locale === "en"
+                  ? pageData.scFragments[1].scContentEn.json[0].content[0].value
+                  : pageData.scFragments[1].scContentFr.json[0].content[0]
+                      .value}
               </h2>
 
-              <div className="row-span-2 bg-gradient-to-tr from-custom-blue-blue to-vc-blue-lt  p-4 ">
-                <img
-                  className=" object-fill w-64 h-auto mx-auto"
-                  src="/virtualconcierge/VirtualConcierge.svg"
-                  alt={t("vc:virtualAssistantLogoAlt")}
+              <div className="row-span-2 bg-gradient-to-tr from-custom-blue-blue to-vc-blue-lt p-4 flex justify-center">
+                <Image
+                  src={`https://www.canada.ca${
+                    props.locale === "en"
+                      ? pageData.scFragments[2].scImageEn._path
+                      : pageData.scFragments[2].scImageFr._path
+                  }`}
+                  alt={
+                    props.locale === "en"
+                      ? pageData.scFragments[2].scImageAltTextEn
+                      : pageData.scFragments[2].scImageAltTextFr
+                  }
+                  width={260}
+                  height={260}
                 />
               </div>
               <p className=" font-body text-lg px-1 lg:px-0 ">
-                {t("vc:virtualAssistantBioBody")}
+                {props.locale === "en"
+                  ? pageData.scFragments[1].scContentEn.json[1].content[0].value
+                  : pageData.scFragments[1].scContentFr.json[1].content[0]
+                      .value}
               </p>
             </div>
           </div>
@@ -176,9 +232,18 @@ export default function Home(props) {
   );
 }
 
-export const getStaticProps = async ({ locale }) => ({
-  props: {
-    locale: locale,
-    ...(await serverSideTranslations(locale, ["common", "vc"])),
-  },
-});
+export const getStaticProps = async ({ locale }) => {
+  // get page data from AEM
+  const res = await queryGraphQL(getVirtualAssistantPage).then((result) => {
+    return result;
+  });
+
+  const data = res.data.sCLabsPageByPath;
+  return {
+    props: {
+      locale: locale,
+      pageData: data,
+      ...(await serverSideTranslations(locale, ["common"])),
+    },
+  };
+};